内部クラスであったDrawPanelManagerクラスを新しいファイルへ切り出した
