+++ src/application/component/system/GameManager.java
@@ -3,7 +3,6 @@ package application.component.system;
 import application.component.map.GameMap;
 import application.component.map.MapFactory;
 import application.component.map.MapFactory.IllegalMapDataException;
-import application.controller.GameController;
 import javafx.scene.layout.Pane;
 
 import java.util.Timer;
@@ -28,6 +27,13 @@ public class GameManager {
     private DrawPanelManager dpm;   // パネル管理
     // TODO Playerクラスのフィールドを作成
 
+    /**
+     * インスタンスの取得
+     *
+     * @param drawPane the draw pane
+     * @param stageNum the stage num
+     * @return the instance
+     */
     public static GameManager getInstance(Pane drawPane, int stageNum) {
         ourInstance.stageNum = stageNum;
         ourInstance.drawPane = drawPane;
@@ -99,7 +105,7 @@ public class GameManager {
     }
 
     /**
-     * GameMapクラスのインスタンスが持つImageViewをdrawPaneに登録する
+     * GameMapクラスのインスタンスが持つGaneObjectoのインスタンスのImageViewをdrawPaneに登録する
      */
     private void inputObjectsToPane() {
         // TODO 実装する
@@ -120,58 +126,6 @@ public class GameManager {
             //== 描画パネル(drawPane)の移動
             //== 無効キャラクターの削除
             //== ゲーム終了判定
-        }
-    }
-
-    /**
-     * 描画パネル管理クラス
-     */
-    public class DrawPanelManager {
-        private Pane drawPanel;
-
-        public DrawPanelManager(Pane drawPanel) {
-            this.drawPanel = drawPanel;
-        }
-
-        /**
-         * マップの反映
-         *
-         * @param gm the gm
-         */
-        public void inputMap(GameMap gm) {
-            drawPane.getChildren().removeAll();  // 既存要素の削除
-            // TODO GameObject実装後、実装
-//            List<GameObject> nodes = gm.getGameObject();
-//            nodes.stream().forEach(n -> { drawPane.getChildren().add(n) });
-        }
-
-        /**
-         * オブジェクトの反映
-         */
-        // TODO GameObject実装後、実装2
-        public void inputGameObject() {
-
-        }
-
-        /**
-         * パネル(の左上端座標)を親ノードのローカル座標上の指定の座標に移動
-         * また、移動時には、パネルが画面全てを覆うように補正を行う。
-         *
-         * @param x x座標
-         * @param y y座標
-         */
-        public void transfer(double x, double y) {
-            double maxX = 0, maxY = 0;  // 上限
-            double minX = drawPane.getWidth() - GameController.getSceneWidth();    // 下限
-            double minY = drawPane.getHeight() - GameController.getSceneHeight();  // 下限
-
-            // 補正
-            if ( x < minX ) { x = minX; } else if ( x > maxX ) { x = maxX; }
-            if ( y < minY ) { y = minY; } else if ( y > maxY ) { y = maxY; }
-
-            // 格納
-            drawPane.setTranslateX(x);
-            drawPane.setTranslateY(y);
         }
     }
